Put together a `net` module
The changes here bind to the node.js `net` module.
The API tries to stay true to the node.js API while still being usable.

There are a couple of places where things are a bit inconsistent.
The naming for `onCloseServer` and `onCloseSocket`,
`onErrorServer` and `onErrorSocket`, and
`localAddressOption` and `localPortOption`.
This might be showing that there is a split waiting to happen.
Maybe we should have `Node.Net`, `Node.Net.Server` and `Node.Net.Socket`?
Maybe it doesn't matter that much?

In any case, this gives us bindings to the `net` module!