Convert all FFI calls to uncurried versions
From feedback,
we can avoid some allocations if we do the uncurrying explicitly.

This cleans up the FFI implementation a bit at the expense of the PS.
That seems like a fair trade-off since PS is easier to maintain than JS.

There are two bugs that got fixed in this change as well.

The first is that we were typing `Lookup` incorrectly.
The `family` is a `Maybe Int`, not a `Maybe String`.
This was silently failing but not being caught for whatever reason.
There is some kind of concurrency issue with all the callbacks.

The second is that for `Effect _`,
`ado` generates different code from `do`.
The difference here means we get the inefficient version of the code.
Rather than use `ado` and suffer the penalty,
we move over to `do` and allow for more efficient code.