Separate out `Socket` to its own module
Per review,
we can drop the namespace on some values if we separate modules.

Having done the move,
it's not clear that the extra level of nesting was worth it.
It was one or two values that needed a suffix.
But now we've got a whole new module and another file of JS to maintain.

Oh well, we're here now.