Separate out `Server` to its own module
This is the second part of movind to modules.
Again, not sure this is worth it as we end up with more code to maintain
and a less clear API overall.
We should have left well-enough alone and had six out of 78 concepts in
this module inconsistent.

Oh well, maybe someone will find a better set of values in the future.